# _tmp/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SakuraFilter API",
    "description": "工业/汽车滤清器产品管理平台 API\r\n\r\n## 模块说明\r\n- **认证**: JWT 登录/刷新/登出/用户管理\r\n- **公开搜索**: 前台产品搜索/详情 (无需认证)\r\n- **后台产品管理**: CRUD (需 admin 角色)\r\n- **字典管理**: 8 类字典 CRUD (需认证)\r\n- **ETL**: 数据导入/状态/进度 (X-Admin-Token)\r\n- **运维**: 健康检查/指标/性能告警\r\n\r\n## 认证方式\r\n1. **JWT Bearer** (推荐): 登录 /api/auth/login 获取 token, 放入 Authorization: Bearer {token}\r\n2. **X-Admin-Token** (ETL/CI 备用): 放入 X-Admin-Token: {token}",
    "contact": {
      "name": "SakuraFilter Team",
      "email": "admin@sakurafilter.dev"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/public/compare": {
      "get": {
        "tags": [
          "PublicCompare"
        ],
        "summary": "批量对比 (公开) - 1-6 个产品\r\nURL: GET /api/public/compare?ids=1,2,3",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/public/images/{key}": {
      "get": {
        "tags": [
          "SakuraFilter.Api"
        ],
        "operationId": "PublicGetImage",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "X-Admin-Token": {
        "type": "apiKey",
        "description": "Day 8.4: dev 静态 token, 从 appsettings.json 的 Auth:DevStaticToken 读",
        "name": "X-Admin-Token",
        "in": "header"
      },
      "Bearer": {
        "type": "http",
        "description": "JWT Bearer token. 格式: Bearer {token} (登录 /api/auth/login 获取)",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "X-Admin-Token": [ ]
    },
    {
      "Bearer": [ ]
    }
  ],
...
}